backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
fpdf2==2.7.5
